add custSnapRes to save snap ro specific folder
