+++ src/tests/pages/JokesPage.test.tsx
